TH2-155. To make user able to save test events in batches or separately

+++ cradle-core/src/main/java/com/exactpro/cradle/messages/StoredMessageId.java
@@ -38,10 +38,10 @@ public class StoredMessageId implements Serializable
 		StoredMessageBatchId batchId = StoredMessageBatchId.fromString(id);
 		
 		String indexPart = id.substring(batchId.toString().length()+StoredMessageBatchId.IDS_DELIMITER.length());
-		int index;
+		long index;
 		try
 		{
-			index = Integer.parseInt(indexPart);
+			index = Long.parseLong(indexPart);
 		}
 		catch (NumberFormatException e)
 		{
